+++ release_notes.md
@@ -2,5 +2,6 @@
 <!-- Please add your release notes in the following format:
 - My change description (#PR)
 -->
+- Making feature flag for worker indexing opt-out instead of opt-in
 **Release sprint:** Sprint 132
 [ [bugs](https://github.com/Azure/azure-functions-host/issues?q=is%3Aissue+milestone%3A%22Functions+Sprint+132%22+label%3Abug+is%3Aclosed) | [features](https://github.com/Azure/azure-functions-host/issues?q=is%3Aissue+milestone%3A%22Functions+Sprint+132%22+label%3Afeature+is%3Aclosed) ]